requirements/windows.txt
aiohttp==3.7.4.post0
altgraph==0.17
asgiref==3.3.4
async-timeout==3.0.1
asyncqt==0.8.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
comtypes==1.1.10
configparser==5.0.2
coverage==5.5
cricket==0.2.5
fbs==0.9.4
future==0.18.2
GPUtil==1.4.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.0.1
macholib==1.14
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
multidict==5.1.0
nose2==0.10.0
numpy==1.19.5
opencv-python==4.5.1.48
pdoc3==0.9.2
pefile==2019.4.18
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.9.0
PyInstaller==3.4
PyQt5==5.9.2
PyQt5-stubs==5.15.2.0
pystache==0.5.4
pywifi==1.1.12
pywin32==300
pywin32-ctypes==0.2.0
requests==2.25.1
rsa==4.7.2
sip==4.19.8
six==1.15.0
smartypants==2.0.1
SoundCard==0.4.1
speedtest-cli==2.1.3
tkreadonly==0.6.1
typing-extensions==3.10.0.0
urllib3==1.26.4
WMI==1.5.1
yarl==1.6.3
zipp==3.4.1
